focus only on Int array indices for now

+++ src/geointerface.jl
@@ -14,34 +14,34 @@ GeoInterphase.geomtype(f::Feature) = GeoInterphase.geomtype(geometry(f))
 # we have to make use of the GeoInterphase fallbacks that call geomtype on the input
 
 GeoInterphase.ncoord(g::Point) = length(g)
-GeoInterphase.getcoord(g::Point, i) = g[i]
+GeoInterphase.getcoord(g::Point, i::Int) = g[i]
 
 GeoInterphase.ncoord(g::LineString) = length(first(g))
 GeoInterphase.npoint(g::LineString) = length(g)
-GeoInterphase.getpoint(g::LineString, i) = Point(g[i])
+GeoInterphase.getpoint(g::LineString, i::Int) = Point(g[i])
 # TODO what to return for length 0 and 1?
 # TODO should this be an approximate equals for floating point?
-GeoInterphase.isclosed(g::LineString, i) = first(g) == last(g)
+GeoInterphase.isclosed(g::LineString, i::Int) = first(g) == last(g)
 
 GeoInterphase.ncoord(g::Polygon) = length(first(first(g)))
 # TODO this should return a "LineString" according to GeoInterphase, but this cannot directly
 # be identified as such, is that a problem?
 GeoInterphase.getexterior(g::Polygon) = LineString(first(g))
 GeoInterphase.nhole(g::Polygon) = length(g) - 1
-GeoInterphase.gethole(g::Polygon, i) = LineString(g[i + 1])
+GeoInterphase.gethole(g::Polygon, i::Int) = LineString(g[i + 1])
 
 GeoInterphase.ncoord(g::MultiPoint) = length(first(g))
 GeoInterphase.npoint(g::MultiPoint) = length(g)
-GeoInterphase.getpoint(g::MultiPoint, i) = Point(g[i])
+GeoInterphase.getpoint(g::MultiPoint, i::Int) = Point(g[i])
 
 GeoInterphase.ncoord(g::MultiLineString) = length(first(first(g)))
 GeoInterphase.nlinestring(g::MultiLineString) = length(g)
-GeoInterphase.getlinestring(g::MultiLineString, i) = LineString(g[i])
+GeoInterphase.getlinestring(g::MultiLineString, i::Int) = LineString(g[i])
 
 GeoInterphase.ncoord(g::MultiPolygon) = length(first(first(first(g))))
 GeoInterphase.npolygon(g::MultiPolygon) = length(g)
-GeoInterphase.getpolygon(g::MultiPolygon, i) = LineString(g[i])
+GeoInterphase.getpolygon(g::MultiPolygon, i::Int) = LineString(g[i])
 
 GeoInterphase.ncoord(g::GeometryCollection) = GeoInterphase.ncoord(first(g))
 GeoInterphase.ngeom(g::GeometryCollection) = length(g)
-GeoInterphase.getgeom(g::GeometryCollection, i) = geometry(g[i])
+GeoInterphase.getgeom(g::GeometryCollection, i::Int) = geometry(g[i])
